[lib/framework] refactor: begin migration of treesitter queries to on directories
Separate out operations on queries from the queries themselves

This (hopefully) increases reuse throughout the code base
