+++ lib/framework/golang/mod.rs
@@ -1,9 +1,6 @@
-mod common;
-#[cfg(test)]
-mod common_test;
 mod delve;
 pub mod gotest;
 #[cfg(test)]
 mod gotest_test;
 pub(crate) mod runnable_meta;
-
+mod treesitter;
