+++ lib/framework/golang/treesitter/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod operations;
+pub(crate) mod queries;
