+++ lib/framework/golang/treesitter/queries/mod.rs
@@ -1,0 +1,1 @@
+pub mod build_tags;
